requirements.txt
flask
flask-cors
pymongo
pillow
torch
transformers
gunicorn
